Remove referral implementation complexity
