+++ src/utils/referralRequest.ts
@@ -1,7 +1,6 @@
 const MESSAGE_TITLE = 'Monarch referral link request';
-const MESSAGE_LINES = {
-  wallet: 'Wallet',
-} as const;
+const WALLET_LINE = 'Wallet: ';
+const MESSAGE_PREFIX = `${MESSAGE_TITLE}\n\n${WALLET_LINE}`;
 const LINE_ENDING_PATTERN = /\r\n?/g;
 
 export interface ReferralCodeRequestMessage {
@@ -9,22 +8,15 @@ export interface ReferralCodeRequestMessage {
 }
 
 export function buildReferralCodeRequestMessage({ wallet }: ReferralCodeRequestMessage) {
-  return [MESSAGE_TITLE, '', `${MESSAGE_LINES.wallet}: ${wallet}`].join('\n');
+  return `${MESSAGE_PREFIX}${wallet}`;
 }
 
 export function parseReferralCodeRequestMessage(message: string): ReferralCodeRequestMessage | null {
-  const lines = message.replace(LINE_ENDING_PATTERN, '\n').split('\n');
-  if (lines[0] !== MESSAGE_TITLE) return null;
+  const normalizedMessage = message.replace(LINE_ENDING_PATTERN, '\n');
+  if (!normalizedMessage.startsWith(MESSAGE_PREFIX)) return null;
 
-  const fields = new Map<string, string>();
-  for (const line of lines.slice(2)) {
-    const separatorIndex = line.indexOf(': ');
-    if (separatorIndex === -1) continue;
-    fields.set(line.slice(0, separatorIndex), line.slice(separatorIndex + 2).trim());
-  }
-
-  const wallet = fields.get(MESSAGE_LINES.wallet);
-  if (!wallet) return null;
+  const wallet = normalizedMessage.slice(MESSAGE_PREFIX.length).trim();
+  if (!wallet || wallet.includes('\n')) return null;
 
   return {
     wallet,
